래퍼런스 자료형 수정

+++ src/main/java/kr/yuhancert/spring/domain/department/repository/DeptMapRepository.java
@@ -3,6 +3,6 @@ package kr.yuhancert.spring.domain.department.repository;
 import kr.yuhancert.spring.domain.department.entity.DeptMap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeptMapRepository extends JpaRepository<DeptMap, Integer> {
+public interface DeptMapRepository extends JpaRepository<DeptMap, Long> {
 
 }
