+++ src/main/java/kr/yuhancert/spring/domain/department/repository/MajorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MajorRepository extends JpaRepository<Major, Integer> {
+public interface MajorRepository extends JpaRepository<Major, Long> {
     Optional<Major> findByMajorName(String majorName);
 }
